Module Bap_wp.Loader

This module is used to compute and load intruction properties in order to handle instructions with unknown semantics. It exposes the slots that contain these properties.

val registers : ( Bap_core_theory.Theory.Semantics.cls, Bap.Std.Var.Set.t ) Bap_core_theory.KB.slot

Slot containing the registers that are referenced in an assembly instruction.