Module Bap_wp.Z3_utils

module Env = Environment

module Constr = Constraint

val tokenize : string -> string list

Splits up an smtlib2 string into a list of tokens that can be parsed.

val build_str : string list -> string

Builds up an smtlib2 string from a list of tokens.

val get_z3_name : Constr.z3_expr Bap.Std.Var.Map.t -> string -> (Bap.Std.Var.t -> string) -> string option

Looks up a Z3 variable's name in the map based off of the name in BIL notation. fmt is used to add prefixes and suffixes to a variable name. For example, init_RDI_orig.

val get_decls_and_symbols : Env.t -> (Z3.FuncDecl.func_decl * Z3.Symbol.symbol) list

get_decls_and_symbols builds from a the var_map in an environment a mapping of all Z3 func_decl to their symbol. This is a helper function for mk_smtlib2

val mk_smtlib2_single : Env.t -> string -> Constr.t

mk_smtlib2_single env smtlib_str takes in a string representing a valid SMT-Lib-2 statement. The variables in the SMT-Lib statements need to appear in the environment. The intended purpose of this function is generating hypothesis and postconditions for single binary analysis

val mk_smtlib2 : Z3.context -> string -> (Z3.FuncDecl.func_decl * Z3.Symbol.symbol) list -> Constr.t